Fix long value rollover problem by using long longs.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10325 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ MMCore/CircularBuffer.h
@@ -72,7 +72,7 @@ private:
    long imageCounter_;
    long insertIndex_;
    long saveIndex_;
-   unsigned int memorySizeMB_;
+   unsigned long memorySizeMB_;
    unsigned int numChannels_;
    unsigned int numSlices_;
    bool overflow_;
